refactor(frontend): organize agents hooks
