+++ desktop/src/components/settings/panes/AgentDefaultsPane.tsx
@@ -8,7 +8,7 @@ import { SettingsCardRow } from "@/components/settings/shared/SettingsCardRow";
 import { SettingsPageHeader } from "@/components/settings/shared/SettingsPageHeader";
 import { ProviderIcon } from "@/components/ui/icons";
 import { SettingsMenu } from "@/components/ui/SettingsMenu";
-import { useAgentCatalog } from "@/hooks/agents/use-agent-catalog";
+import { useAgentCatalog } from "@/hooks/agents/derived/use-agent-catalog";
 import { withUpdatedDefaultModelIdByAgentKind } from "@/lib/domain/agents/model-options";
 import { withUpdatedDefaultSessionModeByAgentKind } from "@/lib/domain/chat/session-controls/session-mode-control";
 import {
