+++ desktop/src/components/settings/panes/AgentsPane.tsx
@@ -2,8 +2,8 @@ import type { AgentSummary } from "@anyharness/sdk";
 import type { ReactNode } from "react";
 import { useNavigate } from "react-router-dom";
 import { AGENTS_PAGE_COPY } from "@/copy/agents/agents-copy";
-import type { AgentsPaneRowState } from "@/hooks/agents/use-agents-pane-state";
-import { useAgentsPaneState } from "@/hooks/agents/use-agents-pane-state";
+import type { AgentsPaneRowState } from "@/hooks/agents/facade/use-agents-pane-state";
+import { useAgentsPaneState } from "@/hooks/agents/facade/use-agents-pane-state";
 import { AgentSetupModal } from "@/components/agents/AgentSetupModal";
 import { Badge } from "@/components/ui/Badge";
 import { Button } from "@/components/ui/Button";
