+++ desktop/src/hooks/agents/cache/query-keys.ts
