+++ desktop/src/hooks/agents/cache/use-local-agent-credentials.ts
