+++ desktop/src/hooks/agents/derived/use-agent-catalog.ts
