+++ desktop/src/hooks/plans/use-plan-handoff-workflow.ts
@@ -6,7 +6,7 @@ import type {
 } from "@anyharness/sdk";
 import { useSetSessionConfigOptionMutation } from "@anyharness/sdk-react";
 import { PLAN_HANDOFF_DEFAULT_PROMPT } from "@/copy/plans/plan-prompts";
-import { useAgentCatalog } from "@/hooks/agents/use-agent-catalog";
+import { useAgentCatalog } from "@/hooks/agents/derived/use-agent-catalog";
 import { useActiveSessionLaunchState } from "@/hooks/chat/use-active-chat-session-selectors";
 import { useChatLaunchCatalog } from "@/hooks/chat/use-chat-launch-catalog";
 import { useConfiguredLaunchReadiness } from "@/hooks/chat/use-configured-launch-readiness";
